Refactor, code review
