+++ data-report-commons/src/test/resources/academicArticle.nt
